fix(nav): prevent cookies error by keeping NavWrapper in server scope (events & sponsor pages)
